+++ test/enhancedSet.test.js
@@ -86,26 +86,32 @@ describe('EnhancedSet unit tests', () => {
     });
   });
 
+  describe('.toArray()', () => {
+    it('converts the set into an array', () => {
+      expect(set1.toArray()).to.deep.equal([1, 2, 3, 4]);
+    });
+  });
+
   describe('.product(set)', () => {
     it('applies cortesian product on two sets with default seprator', () => {
       const product = set1.product(set2);
       expect(product.size).to.equal(16);
-      expect(product.has('1,3')).to.equal(true);
-      expect(product.has('1,4')).to.equal(true);
-      expect(product.has('1,5')).to.equal(true);
-      expect(product.has('1,6')).to.equal(true);
-      expect(product.has('2,3')).to.equal(true);
-      expect(product.has('2,4')).to.equal(true);
-      expect(product.has('2,5')).to.equal(true);
-      expect(product.has('2,6')).to.equal(true);
-      expect(product.has('3,3')).to.equal(true);
-      expect(product.has('3,4')).to.equal(true);
-      expect(product.has('3,5')).to.equal(true);
-      expect(product.has('3,6')).to.equal(true);
-      expect(product.has('4,3')).to.equal(true);
-      expect(product.has('4,4')).to.equal(true);
-      expect(product.has('4,5')).to.equal(true);
-      expect(product.has('4,6')).to.equal(true);
+      expect(product.has('13')).to.equal(true);
+      expect(product.has('14')).to.equal(true);
+      expect(product.has('15')).to.equal(true);
+      expect(product.has('16')).to.equal(true);
+      expect(product.has('23')).to.equal(true);
+      expect(product.has('24')).to.equal(true);
+      expect(product.has('25')).to.equal(true);
+      expect(product.has('26')).to.equal(true);
+      expect(product.has('33')).to.equal(true);
+      expect(product.has('34')).to.equal(true);
+      expect(product.has('35')).to.equal(true);
+      expect(product.has('36')).to.equal(true);
+      expect(product.has('43')).to.equal(true);
+      expect(product.has('44')).to.equal(true);
+      expect(product.has('45')).to.equal(true);
+      expect(product.has('46')).to.equal(true);
     });
 
     it('applies cortesian product on two sets with custom seprator', () => {
@@ -135,9 +141,23 @@ describe('EnhancedSet unit tests', () => {
     });
   });
 
-  describe('.toArray()', () => {
-    it('converts the set into an array', () => {
-      expect(set1.toArray()).to.deep.equal([1, 2, 3, 4]);
+  describe('.power(n)', () => {
+    it('returns empty set when n is 0', () => {
+      expect(set1.power(0).size).to.equal(0);
+    });
+
+    it('applies cartesian product on the set itself n times', () => {
+      const s = new EnhancedSet(['0', '1']);
+      expect(s.power(3).toArray()).to.deep.equal([
+        '000',
+        '001',
+        '010',
+        '011',
+        '100',
+        '101',
+        '110',
+        '111'
+      ]);
     });
   });